+++ src/core/fill.js
@@ -20,6 +20,7 @@ THREE.Bootstrap.registerPlugin('fill', {
         [ three.element, document.documentElement ].filter(is).map(set);
     }
 
+    three.canvas.style.display = 'block'
   },
 
   uninstall: function (three) {
@@ -33,6 +34,8 @@ THREE.Bootstrap.registerPlugin('fill', {
 
       this.applied.map(set);
     }
+
+    three.canvas.style.display = ''
   }
 
 });